requirements.txt
Flask==3.0.3
pandas==1.5.3
scikit-learn==1.1.3
nltk==3.6.3
python-docx==0.8.11
PyMuPDF==1.21.1
werkzeug==3.0.3
